% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{callr_get_objects}
\alias{callr_get_objects}
\title{Private. A function for fetch all global objects from a callr session.}
\usage{
callr_get_objects(session)
}
\arguments{
\item{session}{a callr session}
}
\value{
a list of objects
}
\description{
Private. A function for fetch all global objects from a callr session.
}
