% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compatibility.R
\name{datapackager_object_read}
\alias{datapackager_object_read}
\title{datapackager_object_read}
\usage{
datapackager_object_read(object, path = ".")
}
\arguments{
\item{object}{a singleton character vector of an object name saved by \code{dpr_save} in a previously rendered script.}

\item{path}{The relative path to the data package. The default is the
working directory.}
}
\value{
an environment containing the object
}
\description{
A reimplementation of DataPackageR::datapackager_object_read for
processing scripts from DataPackageR packages. This originaly
required a previous process to export the object before this could
be used. The DPR2 equivalent of that is to simply read the rda
file saved by a previous process.
}
\author{
jmtaylor
}
