% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DPR2-package.R
\docType{package}
\name{DPR2-package}
\alias{DPR2}
\alias{DPR2-package}
\title{DPR2: DPR2 // DataPackageR2}
\description{
Build automation and dataset tracking for R data packages. Based on DataPackageR.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://fredhutch.github.io/DPR2/}
}

}
\author{
\strong{Maintainer}: Jason Taylor \email{jmtaylor@fredhutch.org}

}
\keyword{internal}
