% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{dpr_add_objects}
\alias{dpr_add_objects}
\title{dpr_add_objects}
\usage{
dpr_add_objects(objects, path = ".")
}
\arguments{
\item{objects}{the names of objects defined in processing scripts set to
build.}

\item{path}{path to a DPR2 data package}
}
\description{
Add objects to be built by DPR2.
}
\details{
\code{dpr_add_objects} is mostly provided for backwards compatibility with
DataPackageR. It is recommended to use \code{dpr_save} in the processing scripts
directly to save files to the data directory rather than using
\code{dpr_add_objects}.
}
