% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{dpr_add_scripts}
\alias{dpr_add_scripts}
\title{dpr_add_scripts}
\usage{
dpr_add_scripts(scripts, path = ".")
}
\arguments{
\item{scripts}{the names of the scripts in the processing directory
described in the datapackager.yml to be built when running \code{dpr_build}.
scripts currently being tracked.}

\item{path}{path to a data package}
}
\description{
Add processing scripts to list of scripts rendered when a package is
rendered or built.
}
\details{
This function will add new processing scripts to be rendered. To remove
processing scripts, see \code{?dpr_rm_scripts}. To generate a table of scripts
set to build, see \code{?dpr_processing}.
}
