% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version.R
\name{dpr_check_git}
\alias{dpr_check_git}
\title{dpr_check_git}
\usage{
dpr_check_git(path)
}
\arguments{
\item{path}{path to data package}
}
\value{
boolean
}
\description{
Private. Checks for nessessary git features for using dpr_data_history.
}
\author{
jmtaylor
}
