% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version.R
\name{dpr_checksum_data_objects}
\alias{dpr_checksum_data_objects}
\title{dpr_checksum_data_objects}
\usage{
dpr_checksum_data_objects(path)
}
\arguments{
\item{path}{a path to an rda file}
}
\value{
a string of an md5 checksum of an object loaded from a path
}
\description{
Private. Get md5 checksum of object once loaded from a path.
}
\author{
jmtaylor
}
