% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version.R
\name{dpr_checksum_files}
\alias{dpr_checksum_files}
\title{dpr_checksum_files}
\usage{
dpr_checksum_files(paths)
}
\arguments{
\item{paths}{path of files to hash}
}
\value{
a character string of the md5 hashes of a files loaded into memory.
}
\description{
This function will return an md5 checksum of an in-memory R object.
}
