% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version.R
\name{dpr_checksum_from_version}
\alias{dpr_checksum_from_version}
\title{dpr_checksum_from_version}
\usage{
dpr_checksum_from_version(version, path = ".")
}
\arguments{
\item{version}{a string hash of a data version}

\item{path}{a path to a data package}
}
\value{
a string of a checksum from an object recalled from the git history
}
\description{
Get a checksum of an object from the git history.
}
\author{
jmtaylor
}
