% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compatibility.R
\name{dpr_convert}
\alias{dpr_convert}
\title{dpr_convert}
\usage{
dpr_convert(path = ".", renv_init = TRUE)
}
\arguments{
\item{path}{a path value to init at a specific path, when using the
default the current working directory is used.}

\item{renv_init}{to initialize renv at the newly converted data package}
}
\description{
Convert a repository from DataPackageR to DPR2.
}
\details{
This is a mildly destructive process which convert, adds, and
removes files to take information from a DataPackageR repository
and put it in the DPR2 format. This function also removes those
files that are no longer needed from DataPackageR by DPR2. <specify
specifically what is removed, added, and changed>
}
\author{
jmtaylor
}
