% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{dpr_create}
\alias{dpr_create}
\title{dpr_create}
\usage{
dpr_create(
  path = ".",
  yaml = dpr_yaml_init(),
  desc = dpr_description_init(),
  renv_init = TRUE
)
}
\arguments{
\item{path}{A path to the data package.}

\item{yaml}{A returned list for dpr_yaml_init()}

\item{desc}{A returned list for dpr_description_init()}

\item{renv_init}{Logical; whether to initiate renv (default TRUE)}
}
\description{
Create an empty data package. Package is created with
datapackager.yml and DESCRIPTION files as described by
dpr_yaml_init() and dpr_description_init() function calls.
}
\author{
jmtaylor
}
