% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version.R
\name{dpr_data_checksums}
\alias{dpr_data_checksums}
\title{dpr_data_checksums}
\usage{
dpr_data_checksums(path = ".")
}
\arguments{
\item{path}{path to data package}
}
\value{
a data.frame containing the name of the rda object and the
corresponding checksum
}
\description{
Return table of current in-memory checksums for rda files in the data directory.
}
\author{
jmtaylor
}
