% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version.R
\name{dpr_data_digest}
\alias{dpr_data_digest}
\title{dpr_data_digest}
\usage{
dpr_data_digest(path = ".")
}
\arguments{
\item{path}{path to data package}
}
\value{
a data.frame showing the hashes found in the data digest
}
\description{
Return the current hashes recorded to the data digest directory generated during the last
build.
}
\author{
jmtaylor
}
