% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version.R
\name{dpr_data_hashes}
\alias{dpr_data_hashes}
\title{dpr_data_hashes}
\usage{
dpr_data_hashes(path = ".")
}
\arguments{
\item{path}{path to data package}
}
\value{
a data.frame containing the name of the rda object and the
corresponding hash
}
\description{
Return table of current hashes for rda files in the data directory.
}
\author{
jmtaylor
}
