% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{dpr_data_history}
\alias{dpr_data_history}
\title{dpr_data_history}
\usage{
dpr_data_history(include_checksums = FALSE, path = ".")
}
\arguments{
\item{include_checksums}{a boolean value indicating if checksums should be included in the returned data.frame object; computing checksums is less performant}

\item{path}{path to data package}
}
\value{
a data.frame object
}
\description{
Return the history of all the former and current files in the \code{data} directory.
}
\author{
jmtaylor
}
