% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{dpr_data_versions}
\alias{dpr_data_versions}
\title{dpr_data_versions}
\usage{
dpr_data_versions(path = ".")
}
\arguments{
\item{path}{The full path to the data package. The default is the
working directory.}
}
\value{
data.frame
}
\description{
Retrieve data.frame of data object names and versions rendered to the data directory.
}
\author{
jmtaylor
}
