% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{dpr_description_init}
\alias{dpr_description_init}
\title{dpr_description_init}
\usage{
dpr_description_init(...)
}
\arguments{
\item{...}{Arguments are treated as key value pairs for the
package DESCRIPTION file.}
}
\value{
A list of key:value pairs for generating a DESCRIPTION
file.
}
\description{
A function that where each argument converts or adds to the
DESCRIPTION file's key:value pairs. No arguments will return the
default values. See \code{dpr_description_defaults()} for a list of
default values, or call this function with no arguments.
}
\author{
jmtaylor
}
