% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{dpr_description_load}
\alias{dpr_description_load}
\title{dpr_description_load}
\usage{
dpr_description_load(pkgp)
}
\arguments{
\item{pkgp}{the package path}
}
\value{
a desc object
}
\description{
Private. Load the package DESCRIPTION file into memory.
}
\author{
jmtaylor
}
