% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compatibility.R
\name{dpr_dpr1_yaml_convert}
\alias{dpr_dpr1_yaml_convert}
\title{dpr_dpr1_yaml_convert}
\usage{
dpr_dpr1_yaml_convert(path = ".")
}
\arguments{
\item{path}{The relative path to the data package. The default is the
working directory.}
}
\description{
Replaces a DataPackageR yaml with a DPR2 yaml.
}
\author{
jmtaylor
}
