% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compatibility.R
\name{dpr_dpr1_yaml_load}
\alias{dpr_dpr1_yaml_load}
\title{dpr_yaml_convert}
\usage{
dpr_dpr1_yaml_load(path = ".")
}
\arguments{
\item{path}{The relative path to the data package. The default is the
working directory.}
}
\value{
a list of DPR2 yaml key:values pairs
}
\description{
Converts a DataPackageR yaml file to a DPR2 yaml file. Returns a writeable yaml object.
}
\author{
jmtaylor
}
