% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version.R
\name{dpr_envs_to_checksums}
\alias{dpr_envs_to_checksums}
\title{dpr_envs_to_checksums}
\usage{
dpr_envs_to_checksums(envs)
}
\arguments{
\item{envs}{environments to extract checksums}
}
\value{
singleton character vector
}
\description{
Private. Return a vector of md5 checksums from a list of
environments.
}
\author{
jmtaylor
}
