% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{dpr_get_template}
\alias{dpr_get_template}
\title{dpr_get_template}
\usage{
dpr_get_template(regex)
}
\arguments{
\item{regex}{a regular expression}
}
\value{
a character vector
}
\description{
Private. Return a full path of file from installed templates directory.
}
\author{
jmtaylor
}
