% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{dpr_hash_file}
\alias{dpr_hash_file}
\title{dpr_hash_file}
\usage{
dpr_hash_file(path)
}
\arguments{
\item{path}{path of file to hash}
}
\value{
a character string of the SHA1 hash of the file using git style hash padding.
}
\description{
Return hash of file at the path provided. Hash is equivalent to the git blob hash of the file.
}
\author{
jmtaylor
}
