% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version.R
\name{dpr_hash_files}
\alias{dpr_hash_files}
\title{dpr_hash_files}
\usage{
dpr_hash_files(paths)
}
\arguments{
\item{paths}{path of files to hash}
}
\value{
a character string of the md5 hashes of a files loaded into memory.
}
\description{
This function will return an md5 hash of an R object into memory.
}
