% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version.R
\name{dpr_hashes_to_checksums}
\alias{dpr_hashes_to_checksums}
\title{dpr_hashes_to_checksums}
\usage{
dpr_hashes_to_checksums(hashes, path)
}
\arguments{
\item{hashes}{a character vector of full sha1 hashes}

\item{path}{a path to a data package}
}
\value{
a character vector of checksums
}
\description{
Private. From a vector of hashes load the data and generate the
checksum for the objects in memory.
}
\author{
jmtaylor
}
