% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version.R
\name{dpr_hashes_to_envs}
\alias{dpr_hashes_to_envs}
\title{dpr_git_hashes_to_envs}
\usage{
dpr_hashes_to_envs(hashes, path = ".")
}
\arguments{
\item{hashes}{a character vector of full sha1 hashes}

\item{path}{a path to a data package}
}
\value{
list of envs containing objects loaded from git history
}
\description{
Private. Pulls rda blobs from the git objects database and returns
them loaded in environments with empty env as parent.
}
\author{
jmtaylor
}
