% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{dpr_init}
\alias{dpr_init}
\title{dpr_init}
\usage{
dpr_init(path = ".")
}
\description{
A wrapper for dpr_create that uses the parent directory as the
location to create the data package, setting the package name as th
}
\author{
jmtaylor
}
