% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{dpr_is}
\alias{dpr_is}
\title{dpr_is}
\usage{
dpr_is(path)
}
\arguments{
\item{path}{path to datapackage}
}
\value{
boolean
}
\description{
Private. Checks if path is a DPR2 datapackage or not.
}
\author{
jmtaylor
}
