% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compatibility.R
\name{dpr_is_dpr1}
\alias{dpr_is_dpr1}
\title{dpr_is_dpr1}
\usage{
dpr_is_dpr1(path = ".")
}
\arguments{
\item{path}{The relative path to the data package. The default is the working directory.}
}
\value{
a boolean value indicating whether the specified directory contains a DataPackageR package or not
}
\description{
A private function to verify whether a specified directory contains a DataPackageR package.
}
\author{
jmtaylor
}
