% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compatibility.R
\name{dpr_is_dpr2}
\alias{dpr_is_dpr2}
\title{dpr_is_dpr2}
\usage{
dpr_is_dpr2(path = ".")
}
\arguments{
\item{path}{path to datapackage}
}
\value{
logical
}
\description{
A private function to verify whether a specified directory contains a DPR2 package or not
}
\author{
jmtaylor
}
