% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version.R
\name{dpr_list_rda}
\alias{dpr_list_rda}
\title{dpr_list_rda}
\usage{
dpr_list_rda(path)
}
\arguments{
\item{path}{path to data package}
}
\value{
a character vector of rda files in the data directory
}
\description{
Private. Standard way of listing rda files available in data for use across all functions.
}
\author{
jmtaylor
}
