% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{dpr_objects}
\alias{dpr_objects}
\title{dpr_objects}
\usage{
dpr_objects(path = ".")
}
\arguments{
\item{path}{path to a DPR2 data package}
}
\value{
a data frame displaying object names in the data directory, and if
they are tracked or not
}
\description{
Return a data frame of obejcts currently being tracked.
}
\details{
The data frame returned by \verb{dpr_objects displays what objects will be saved from the processing environments to the }data\verb{directory and reports what objects are tracked by DPR2. Objects are added to this list by using}dpr_track_objects\code{. }dpr_track_objects\verb{is mostly provided for backwards compatibility with DataPackageR. It is recommended to use}dpr_save\verb{in the processing scripts directly to save files to the data directory rather than using}dpr_track_objects`.
}
