% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{dpr_path}
\alias{dpr_path}
\title{Construct path from data package directory}
\usage{
dpr_path(...)
}
\arguments{
\item{...}{Trailing path components passed to \code{\link[=file.path]{file.path()}}. All arguments must
be the same length or length one.}
}
\value{
The normalized path with the additional path components appended.
Throws an error if no root is found.
}
\description{
Construct path from data package directory
}
