% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{dpr_processes}
\alias{dpr_processes}
\title{dpr_processes}
\usage{
dpr_processes(path = ".")
}
\arguments{
\item{path}{path to a DPR2 data package}
}
\value{
a data frame displaying process names in the processing directory,
and if they are tracked or not
}
\description{
Return a data frame of processes currently being tracked.
}
\details{
The data frame returned by \code{dpr_processes} displays what processing scripts
are found in the processing directory (see, ?dpr_yaml_get) and which of those
will be rendered by calling \code{dpr_render} or \code{dpr_build}.
}
