% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{dpr_purge_data_directory}
\alias{dpr_purge_data_directory}
\title{dpr_purge_data_directory}
\usage{
dpr_purge_data_directory(path = ".", yml)
}
\arguments{
\item{path}{the package path}

\item{yml}{an R yaml list object}
}
\value{
nothing
}
\description{
Private. A function for purging the data directory.
}
\author{
jmtaylor
}
