% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{dpr_recall_data_version}
\alias{dpr_recall_data_version}
\title{dpr_recall_data_version}
\usage{
dpr_recall_data_version(version, path = ".")
}
\arguments{
\item{version}{the hash of the data to recall; partial hashes allowed from 1 to n digits}

\item{path}{the path to the data package}
}
\value{
a list of objects loaded from an rda file pulled from the git history
}
\description{
Load a data version into memory by its hash. Only rda files in the data directory will be recalled.
}
\author{
jmtaylor
}
