% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{dpr_render}
\alias{dpr_render}
\title{dpr_render}
\usage{
dpr_render(path = ".", ...)
}
\arguments{
\item{path}{The relative path to the data package. The default is the
working directory.}

\item{...}{datapakager.yml value overrides. When arguments are
specified, those arguments are used as the YAML key value pairs
instead of what is specified by the \code{datapackager.yml}.}
}
\value{
does not return anything but performs rendering, processing and
data-saving operation defined in configuration file
}
\description{
The dpr_render function process and render all processing scripts defined in the
datapackager.yml configuration file. Does  not build or install the data package.
For full package build and installation, use the dpr_build function.
}
\author{
jmtaylor
}
