% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{dpr_rm_objects}
\alias{dpr_rm_objects}
\title{dpr_rm_objects}
\usage{
dpr_rm_objects(objects, path = ".")
}
\arguments{
\item{objects}{a character vector of object names to be tracked, addign them
to the data directory after a processing script is rendered.}

\item{path}{path to a DPR2 data package}
}
\description{
Remove objects to be tracked by DPR2.
}
\details{
\code{dpr_track_objects} is mostly provided for backwards compatibility with
DataPackageR. It is recommended to use \code{dpr_save} in the processing scripts
directly to save files to the data directory rather than using
\code{dpr_track_objects}.
}
