% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{dpr_rm_processes}
\alias{dpr_rm_processes}
\title{dpr_rm_processes}
\usage{
dpr_rm_processes(processes, path = ".")
}
\arguments{
\item{processes}{names of processes, found in the processing_directory set in
the \code{datapackager.yml} file, to be removed from tracking and no long run
when a script is rendered.}

\item{path}{path to a DPR2 data package}
}
\description{
Remove processing scripts from list of scripts rendered when a package is
rendered or built.
}
\details{
This function will remove processing scripts from those that will be
rendered. To add processing scripts, see \code{?dpr_track_processes}. To generate
a table of processes currently being tracked, see \code{?dpr_processing}.
}
