% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{dpr_rm_scripts}
\alias{dpr_rm_scripts}
\title{dpr_rm_scripts}
\usage{
dpr_rm_scripts(scripts, path = ".")
}
\arguments{
\item{scripts}{names of scripts, found in the processing_directory set in
the \code{datapackager.yml} file, to be removed from to_build and no long run
when a script is rendered.}

\item{path}{path to a DPR2 data package}
}
\description{
Remove processing scripts from list of scripts rendered when a package is
rendered or built.
}
\details{
This function will remove processing scripts from those that will be
rendered. To add processing scripts, see \code{?dpr_add_scripts}. To generate
a table of scripts currently being tracked, see \code{?dpr_scripts}.
}
