% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{dpr_scripts}
\alias{dpr_scripts}
\title{dpr_scripts}
\usage{
dpr_scripts(path = ".")
}
\arguments{
\item{path}{path to a DPR2 data package}
}
\value{
a data frame displaying script names in the processing directory,
and if they are set to build or not
}
\description{
Return a data frame of scripts currently being tracked.
}
\details{
The data frame returned by \code{dpr_scripts} displays what processing scripts
are found in the processing directory (see, ?dpr_yaml_get) and which of those
will be rendered by calling \code{dpr_render} or \code{dpr_build}.
}
