% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{dpr_set_keys}
\alias{dpr_set_keys}
\title{dpr_set_keys}
\usage{
dpr_set_keys(old, new)
}
\arguments{
\item{old}{key:values list object}

\item{new}{key:values list object to add/modify}
}
\value{
list
}
\description{
Private. Replace and add key:value pairs in old list with new.
}
\author{
jmtaylor
}
