% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{dpr_track_objects}
\alias{dpr_track_objects}
\title{dpr_track_objects}
\usage{
dpr_track_objects(objects, path = ".")
}
\arguments{
\item{objects}{the names of instatiated objects defined in tracked processing
scripts.}

\item{path}{path to a DPR2 data package}
}
\description{
Add objects to be tracked by DPR2.
}
\details{
\code{dpr_track_objects} is mostly provided for backwards compatibility with
DataPackageR. It is recommended to use \code{dpr_save} in the processing scripts
directly to save files to the data directory rather than using
\code{dpr_track_objects}.
}
