% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{dpr_track_processes}
\alias{dpr_track_processes}
\title{dpr_track_processes}
\usage{
dpr_track_processes(processes, path = ".")
}
\arguments{
\item{processes}{the names of the processes in the processing directory
described in the datapackager.yml to be built when running \code{dpr_build}.
processes currently being tracked.}

\item{path}{path to a data package}
}
\description{
Add processing scripts to list of scripts rendered when a package is
rendered or built.
}
\details{
This function will add new processing scripts to be rendered. To remove
processing scripts, see \code{?dpr_rm_processes}. To generate a table of processes
currently being tracked, see \code{?dpr_processing}.
}
