% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version.R
\name{dpr_update_data_digest}
\alias{dpr_update_data_digest}
\title{dpr_update_data_digest}
\usage{
dpr_update_data_digest(path = ".", yml)
}
\arguments{
\item{path}{path to data package}

\item{yml}{the yml list object used for the build}
}
\value{
nothing
}
\description{
Private. A function for updating the data digest components by generating hashes
for all .rda files in the specified directory
}
\author{
jmtaylor
}
