% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version.R
\name{dpr_validate_data_digest_source}
\alias{dpr_validate_data_digest_source}
\title{dpr_validate_data_digest_source}
\usage{
dpr_validate_data_digest_source(path)
}
\arguments{
\item{path}{path to data package}
}
\value{
a list of a data digest source
}
\description{
Private. Reads a data digest source, validates it's contents, and returns a list.
}
\author{
jmtaylor
}
