% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{dpr_yaml_get}
\alias{dpr_yaml_get}
\title{dpr_yaml_get}
\usage{
dpr_yaml_get(path = ".", ...)
}
\arguments{
\item{path}{The full path to the data package. The default is the
working directory.}

\item{...}{datapakager.yml value overrides. When arguments are
specified, those arguments are used as the YAML key:value pairs
instead of what is specified by the \code{datapackager.yml}.}
}
\value{
list
}
\description{
Get the current datapackager.yml key:value pairs and add to or
modify those with with new values. All additions and modifications
happen only in memory, not on disc.
}
\author{
jmtaylor
}
