% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{dpr_yaml_init}
\alias{dpr_yaml_init}
\title{dpr_yaml_init}
\usage{
dpr_yaml_init(...)
}
\arguments{
\item{...}{Arguments are treated as key value pairs for the
datapackager.yml file.}
}
\value{
A list of key:value pairs for generating a
datapackager.yml file.
}
\description{
A function where each argument ammends or adds to the
datapackager.yml file's key:value pairs. No arguments will return the
datapackager.yml default values. see \code{dpr_yaml_defaults()} for a list of
default values, or call this function with no arguments.
}
\author{
jmtaylor
}
