% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{dpr_yaml_init_set}
\alias{dpr_yaml_init_set}
\title{dpr_yaml_init_set}
\usage{
dpr_yaml_init_set(yml, pkgp)
}
\arguments{
\item{yml}{an R yaml object.}

\item{pkgp}{the package path}
}
\description{
Private. A function that sets datapackager.yml file
key:values pairs in a new data package.
}
\author{
jmtaylor
}
