% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{dpr_yaml_load}
\alias{dpr_yaml_load}
\title{dpr_yaml_load}
\usage{
dpr_yaml_load(path = ".")
}
\arguments{
\item{path}{the package path}
}
\value{
a yaml object
}
\description{
Private. Load \code{datapackager.yml} into memory.
}
\author{
jmtaylor
}
