% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{dpr_yaml_set}
\alias{dpr_yaml_set}
\title{dpr_yaml_set}
\usage{
dpr_yaml_set(path = ".", ...)
}
\arguments{
\item{path}{The full path to the data package. The default is the
working directory.}

\item{...}{datapakager.yml value overrides. When arguments are
specified, those arguments are used as the YAML key value pairs
instead of what is specified by the \code{datapackager.yml}.}
}
\value{
nothing
}
\description{
Write new datapackager.yml with new or modified key:value pairs.
}
\author{
jmtaylor
}
