% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{dpr_yaml_validate}
\alias{dpr_yaml_validate}
\title{dpr_yaml_validate}
\usage{
dpr_yaml_validate(yml)
}
\arguments{
\item{yml}{a parsed yaml object}
}
\description{
Private. Check yaml values that must be specific.
}
\author{
jmtaylor
}
