% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{dpr_yaml_value_check}
\alias{dpr_yaml_value_check}
\title{dpr_yaml_value_check}
\usage{
dpr_yaml_value_check(yml)
}
\arguments{
\item{yml}{a yaml object or list}
}
\value{
elements of the list of key value pairs to check that did not pass
}
\description{
Private. Check that some yaml keys have specific values.
}
\author{
jmtaylor
}
