% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{dpr_build}
\alias{dpr_build}
\title{dpr_build}
\usage{
dpr_build(path = ".", ...)
}
\arguments{
\item{path}{The relative path to the data package. The default is the
working directory.}

\item{...}{datapackager.yml value overrides. When arguments are specified,
those arguments are used as the YAML key value pairs instead of what is
specified by the \code{datapackager.yml}.}
}
\value{
does not return any value. It performs rendering, building and
installing the package based on the configuration file
}
\description{
Process, render and build data package.
}
\author{
jmtaylor
}
