% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{dpr_build}
\alias{dpr_build}
\title{dpr_build}
\usage{
dpr_build(path = ".", ...)
}
\arguments{
\item{path}{The full path to the data package. The default is the
working directory.}

\item{...}{datapakager.yml value overrides. When arguments are
specified, those arguments are used as the YAML key value pairs
instead of what is specified by the \code{datapackager.yml}.}
}
\value{
nothing
}
\description{
Render and build data package. Uses a special environment,
\code{dpr_build_env}, for the evaluation environment. \code{dpr_build_env}
is removed from the .GlobalEnv once complete.
}
\author{
jmtaylor
}
