% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version.R
\name{dpr_check_git}
\alias{dpr_check_git}
\title{dpr_check_git}
\usage{
dpr_check_git(path)
}
\arguments{
\item{path}{path to data package}
}
\value{
a logical value indicating if the conditions are satisfied
}
\description{
Private. Verifies if necessary conditions are met for accessing Git features
}
\author{
jmtaylor
}
