% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version.R
\name{dpr_compare_data_digest}
\alias{dpr_compare_data_digest}
\title{dpr_compare_data_digest}
\usage{
dpr_compare_data_digest(path = ".", display_digits = 7)
}
\arguments{
\item{path}{path to data package}

\item{display_digits}{number of characters to display for hash values}
}
\value{
a data.frame with the file name, corresponding data hash, existing data_digest hash and a boolean
value indicating if they are same or not
}
\description{
A data digest comparison table, comparing the current data digest
hashes with the current file hashes in the data directory.
}
\details{
Compare the current hashes for the data listed in the data
directory with the hashes listed in the data digest. The data
digest is only updated when the package is built, not when it is
simply rendered using \code{dpr_render()}.
}
\author{
jmtaylor
}
