% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version.R
\name{dpr_compare_data_digest}
\alias{dpr_compare_data_digest}
\title{dpr_compare_data_digest}
\usage{
dpr_compare_data_digest(path = ".")
}
\arguments{
\item{path}{path to data package}
}
\value{
a data.frame with the file name, corresponding data in-memory
checksum, existing data_digest checksum, and a boolean value indicating if
they are same or not
}
\description{
A data digest comparison table, comparing the current data digest checksums
with the current file checksums in the data directory.
}
\details{
Compare the current checksums for the data listed in the data directory with
the checksums listed in the data digest. The data digest is only updated when
the package is built, not when it is simply rendered using \code{dpr_render()}.
}
\author{
jmtaylor
}
