% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compatibility.R
\name{dpr_convert}
\alias{dpr_convert}
\title{dpr_convert}
\usage{
dpr_convert(path = ".")
}
\arguments{
\item{path}{a path value to init at a specific path, when using the
default the current working directory is used.}
}
\description{
Convert a repository from DataPackageR to DPR2.
}
\details{
This is a mildly destructive process which converts, adds, and
removes files to take information from a DataPackageR repository
and put it in the DPR2 format. This function also removes those
files that are no longer needed from DataPackageR by DPR2. This
includes files \code{DATADIGEST}, \code{NEWS.md}, \code{Read-and-delete-me},
\code{R/documentation.R}, and \code{data-raw/documentation.R}, in addition to
the folder \code{inst/extdata/Logfiles}, and any file in \code{inst/doc} with
the same name as an \code{R} or \code{Rmd} file found in the \code{data-raw}
or\code{Logfiles} folders. Modifications to the package \code{DESCRIPTION}
file are also made, removing its \code{Date}, and \code{DataVersion}
fields.
}
