% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{dpr_create}
\alias{dpr_create}
\title{dpr_create}
\usage{
dpr_create(
  path = ".",
  yaml = dpr_yaml_init(),
  desc = dpr_description_init(),
  renv_init = TRUE
)
}
\arguments{
\item{path}{A path to the data package.}

\item{yaml}{A returned list for \code{\link[=dpr_yaml_init]{dpr_yaml_init()}}}

\item{desc}{A returned list for \code{\link[=dpr_description_init]{dpr_description_init()}}}

\item{renv_init}{Logical; whether to initiate renv (default TRUE)}
}
\description{
Initialize a data package by creating a structured directory skeleton for a
new R package. Sets up essential directories, initializes package metadata
in datapackager.yml and DESCRIPTION files as described by \code{\link[=dpr_yaml_init]{dpr_yaml_init()}}
and \code{\link[=dpr_description_init]{dpr_description_init()}} function calls.
}
\author{
jmtaylor
}
