% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version.R
\name{dpr_data_history}
\alias{dpr_data_history}
\title{dpr_data_history}
\usage{
dpr_data_history(path = ".", include_checksums = FALSE)
}
\arguments{
\item{path}{path to data package}

\item{include_checksums}{a boolean value indicating if checksums should be
included in the returned data.frame object; computing checksums is less
performant}
}
\value{
a data.frame object with the git hash, file name, author name, time
of creation and md5 checksum of the file
}
\description{
Return the history of all the former and current files in the \code{data}
directory.
}
