% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{dpr_description_defaults}
\alias{dpr_description_defaults}
\title{dpr_description_defaults}
\usage{
dpr_description_defaults()
}
\value{
a list
}
\description{
Default key:value pairs for the data package DESCRIPTION file.
}
\author{
jmtaylor
}
