% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{dpr_description_init}
\alias{dpr_description_init}
\title{dpr_description_init}
\usage{
dpr_description_init(...)
}
\arguments{
\item{...}{Arguments are treated as key value pairs for the package
\code{DESCRIPTION} file.}
}
\value{
A list of key:value pairs for generating a \code{DESCRIPTION} file.
}
\description{
A function that provides a convenient way to initialize a package's
\code{DESCRIPTION} file by merging default settings with user specific values to the
\code{DESCRIPTION} file's key:value pairs. If no arguments are provided, it will
return the default values. See \code{dpr_description_defaults()} for a list of
default values, or call this function with no arguments.
}
\author{
jmtaylor
}
