% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{dpr_description_init_set}
\alias{dpr_description_init_set}
\title{dpr_description_init_set}
\usage{
dpr_description_init_set(desc, pkgp)
}
\arguments{
\item{desc}{an R desc object.}

\item{pkgp}{the package path}
}
\description{
Private. A function that generates sets DESCRIPTION file
key:values pairs in a new data package.
}
\author{
jmtaylor
}
