% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{dpr_description_set}
\alias{dpr_description_set}
\title{dpr_description_set}
\usage{
dpr_description_set(path = ".", ...)
}
\arguments{
\item{path}{The full path to the data package. The default is the working
directory.}

\item{...}{datapackager.yml value overrides. When arguments are specified,
those arguments are used as the YAML key value pairs instead of what is
specified by the \code{datapackager.yml}. See \code{?dpr_description_defaults} for
the default values that can be set.}
}
\value{
nothing
}
\description{
Write new \code{DESCRIPTION} file with new or modified key:value pairs.
}
