% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version.R
\name{dpr_hash_file}
\alias{dpr_hash_file}
\title{dpr_hash_file}
\usage{
dpr_hash_file(path)
}
\arguments{
\item{path}{path of file to hash}
}
\value{
a character string of the SHA1 hash of the file using git style hash padding.
}
\description{
Return hash of file at the path provided. Hash is equivalent to the git blob hash of the file.
}
\details{
This function will return a git style sha1 hash of a file found at the path argument. Git style
hashing pads the front of the file with "blob" the length of the file in bytes, a null
character, and then the binary content of the file itself. This style of hashing is used here to
ensure that data objects created across objects that have the same blob can be compared whether
the package is under version control or not, or packages that are put under version control at
future dates can compare objects with versions of the package from before version control was
applied.
}
\author{
jmtaylor
}
