% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{dpr_init}
\alias{dpr_init}
\title{dpr_init}
\usage{
dpr_init(path = ".")
}
\arguments{
\item{path}{a path value to init at a specific path, when using
the default the current workig directory is used.}
}
\description{
A wrapper for dpr_create that uses the parent directory as the
location to create the data package, setting the package name as
the current directory.
}
\author{
jmtaylor
}
