% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{dpr_init}
\alias{dpr_init}
\title{dpr_init}
\usage{
dpr_init(
  path = ".",
  yaml = dpr_yaml_init(),
  desc = dpr_description_init(Package = basename(path))
)
}
\arguments{
\item{path}{a path value to init at a specific path. The default
path is the current working directory.}

\item{yaml}{A returned list from \code{\link[=dpr_yaml_init]{dpr_yaml_init()}}}

\item{desc}{A returned list from \code{\link[=dpr_description_init]{dpr_description_init()}}. The default
argument sets the package name to the name of the directory containing the
data package.}
}
\description{
A wrapper for \code{dpr_create} that uses the parent directory as the location to
create the data package, setting the package name as the current directory.
}
\author{
jmtaylor
}
