% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{dpr_init}
\alias{dpr_init}
\title{dpr_init}
\usage{
dpr_init(
  path = ".",
  yaml = dpr_yaml_init(),
  desc = dpr_description_init(),
  renv_init = TRUE
)
}
\arguments{
\item{path}{a path value to init at a specific path, when using
the default the current working directory is used.}

\item{yaml}{A returned list for dpr_yaml_init()}

\item{desc}{A returned list for dpr_description_init()}

\item{renv_init}{Logical; whether to initiate renv (default TRUE)}
}
\description{
A wrapper for dpr_create that uses the parent directory as the
location to create the data package, setting the package name as
the current directory.
}
\author{
jmtaylor
}
