% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compatibility.R
\name{dpr_is_dpr1}
\alias{dpr_is_dpr1}
\title{dpr_is_dpr1}
\usage{
dpr_is_dpr1(path = ".")
}
\arguments{
\item{path}{The relative path to the data package. The default is the working directory.}
}
\value{
logical
}
\description{
Private. Checks if path is a DataPackageR package.
}
\author{
jmtaylor
}
