% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{dpr_objects}
\alias{dpr_objects}
\title{dpr_objects}
\usage{
dpr_objects(path = ".")
}
\arguments{
\item{path}{path to a DPR2 data package}
}
\value{
a data frame displaying object names in the data directory, and if
they are set to build or not
}
\description{
Return a data frame of obejcts currently set to build.
}
\details{
The data frame returned by \code{dpr_objects} displays what objects will be saved
from the processing environments to the \code{data} directory and reports what
objects are set to be built by DPR2. Objects are added to this list by using
\code{dpr_add_objects}. \code{dpr_add_objects} is mostly provided for backwards
compatibility with DataPackageR. It is recommended to use \code{dpr_save} in the
processing scripts directly to save files to the data directory rather than
using \code{dpr_add_objects}.
}
