% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version.R
\name{dpr_recall_data_versions}
\alias{dpr_recall_data_versions}
\title{dpr_recall_data_version}
\usage{
dpr_recall_data_versions(hashes, path = ".")
}
\arguments{
\item{hashes}{the git hashes of the data to recall; partial hashes
allowed from 1 to 40 hexadecimal digits}

\item{path}{the path to the data package}
}
\value{
a list with names of corresponding .rda object and the
versions generated from the object
}
\description{
Load a data version into memory by its hash. Only RDA files in the
data directory will be recalled.
}
\author{
jmtaylor
}
