% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{dpr_render}
\alias{dpr_render}
\title{dpr_render}
\usage{
dpr_render(path = ".", ...)
}
\arguments{
\item{path}{The full path to the data package. The default is the
working directory.}

\item{...}{datapakager.yml value overrides. When arguments are
specified, those arguments are used as the YAML key value pairs
instead of what is specified by the \code{datapackager.yml}.}
}
\value{
nothing
}
\description{
Render all processing scripts in the for the data package. Does
not build. Using the \code{dpr_render()} calling environment. For
evaluation.
}
\author{
jmtaylor
}
