% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{dpr_save}
\alias{dpr_save}
\title{dpr_save}
\usage{
dpr_save(objects, path = dpr_path(), envir = parent.frame())
}
\arguments{
\item{objects}{Character vector of object names to be saved from the
environment specified in \code{envir}.}

\item{path}{The path to the data package. Defaults to \code{\link[=dpr_path]{dpr_path()}}.}

\item{envir}{The environment to search for objects to save. Defaults to
calling environment.}
}
\value{
The original \code{objects} argument, invisibly.
}
\description{
A convenience function for writing data objects to the package data
directory.
}
\author{
jmtaylor
}
