% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{dpr_save}
\alias{dpr_save}
\title{dpr_save}
\usage{
dpr_save(objects, path = ".", envir = parent.frame())
}
\arguments{
\item{objects}{a character vector of object names to saved from the calling environment}

\item{path}{The relative path to the data package. The default is the working directory.}

\item{envir}{The environment to search for objects to save. Defaults to calling environment.}
}
\description{
A convenience function for writing data objects to the package data directory.
}
\author{
jmtaylor
}
