% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{dpr_save}
\alias{dpr_save}
\title{dpr_save}
\usage{
dpr_save(object)
}
\arguments{
\item{object}{an object to save to the path set for the yaml data_directory value}
}
\description{
A convenience function for writing data objects to the package data directory.
}
\author{
jmtaylor
}
