% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{dpr_update_data_digest}
\alias{dpr_update_data_digest}
\title{dpr_update_data_digest}
\usage{
dpr_update_data_digest(path = ".", yml)
}
\arguments{
\item{path}{the package path}

\item{yml}{an R yaml list object}
}
\value{
nothing
}
\description{
Private. A function for updating the data digest components.
}
\author{
jmtaylor
}
