% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{dpr_yaml_defaults}
\alias{dpr_yaml_defaults}
\title{dpr_yaml_defaults}
\usage{
dpr_yaml_defaults()
}
\value{
a list
}
\description{
Default key:value pairs for the data package datapackager.yml.
}
\details{
Definitions of yaml key:value pairs
}
\author{
jmtaylor
}
