% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{dpr_yaml_load}
\alias{dpr_yaml_load}
\title{dpr_yaml_load}
\usage{
dpr_yaml_load(pkgp)
}
\arguments{
\item{pkgp}{the package path}
}
\value{
a yaml object
}
\description{
Private. Load \code{datapackager.yml} into memory. If missing, throw exception.
}
\author{
jmtaylor
}
