% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compatibility.R
\name{project_extdata_path}
\alias{project_extdata_path}
\title{Construct path from data package inst/extdata directory (deprecated)}
\usage{
project_extdata_path(...)
}
\arguments{
\item{...}{Trailing path components passed to \code{\link[=file.path]{file.path()}}. All arguments
must be the same length or length one.}
}
\value{
The normalized path with the additional path components appended.
Throws an error if no root is found.
}
\description{
This function is a compatibility wrapper for legacy code and may be removed
in the future. Please use \code{\link[=dpr_path]{dpr_path()}} instead.
}
